multi-modal-transformers/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-modal-transformers"
version = "0.1.0"
description = "Multi-Modal Transformer Architecture with Custom Attention Mechanisms"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "LlamaSearch.AI Team", email = "info@llamasearch.ai"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.13.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.2.0",
    "wandb>=0.15.0",
    "tensorboard>=2.12.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "safetensors>=0.3.0",
    "huggingface-hub>=0.16.0",
    "albumentations>=1.3.0",
    "pycocotools>=2.0.6",
    "sentencepiece>=0.1.99",
    "accelerate>=0.20.0",
    "lightning>=2.0.0",
    "hydra-core>=1.3.0",
    "opencv-python>=4.7.0",
    "jsonlines>=3.1.0",
    "spacy>=3.5.0",
    "fairseq>=0.12.0",
    "mediapipe>=0.10.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "gradio>=3.40.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.3.2",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "nbsphinx>=0.9.0",
    "mkdocs>=1.4.3",
    "mkdocs-material>=9.1.15",
]

[project.urls]
"Homepage" = "https://github.com/llamasearch/multi-modal-transformers"
"Bug Tracker" = "https://github.com/llamasearch/multi-modal-transformers/issues"
"Documentation" = "https://multi-modal-transformers.llamasearch.ai"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*" 